Fixed data type

+++ src/harnesses/harnesses.py
@@ -139,7 +139,7 @@ class LinearScalarizationHarness(SimpleHarness):
         # Assuming that we have MultiObjectiveLoss here
         preference_ray = get_reference_directions('das-dennis', len(self.cfg.model.loss_fn),
                                                   n_partitions=self.cfg.train.num_preference_vector_partitions)
-        preference_ray = torch.from_numpy(preference_ray[self.cfg.train.preference_ray_idx])
+        preference_ray = torch.from_numpy(preference_ray[self.cfg.train.preference_ray_idx]).to(dtype=torch.float)
         # And assuming we have linear scalarization trainer here
         trainer = self.cfg.train.trainer(model, train_loader, optim, preference_ray)
         checkpointer = DetectionCheckpointer(
